pyproject.toml
[project]
name = "history1"
version = "0.1.0"
description = "Сайт с карточками исторических деятелей"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.11",
]
